+++ crates/collab/src/db/tables/worktree_diagnostic_summary.rs
@@ -1,18 +1,18 @@
-use super::ProjectId;
+use crate::db::ProjectId;
 use sea_orm::entity::prelude::*;
 
 #[derive(Clone, Debug, PartialEq, Eq, DeriveEntityModel)]
-#[sea_orm(table_name = "worktree_repositories")]
+#[sea_orm(table_name = "worktree_diagnostic_summaries")]
 pub struct Model {
     #[sea_orm(primary_key)]
     pub project_id: ProjectId,
     #[sea_orm(primary_key)]
     pub worktree_id: i64,
     #[sea_orm(primary_key)]
-    pub work_directory_id: i64,
-    pub scan_id: i64,
-    pub branch: Option<String>,
-    pub is_deleted: bool,
+    pub path: String,
+    pub language_server_id: i64,
+    pub error_count: i32,
+    pub warning_count: i32,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
