+++ crates/collab/src/db/tables/worktree_entry.rs
@@ -1,18 +1,26 @@
-use super::ProjectId;
+use crate::db::ProjectId;
 use sea_orm::entity::prelude::*;
 
 #[derive(Clone, Debug, PartialEq, Eq, DeriveEntityModel)]
-#[sea_orm(table_name = "worktree_diagnostic_summaries")]
+#[sea_orm(table_name = "worktree_entries")]
 pub struct Model {
     #[sea_orm(primary_key)]
     pub project_id: ProjectId,
     #[sea_orm(primary_key)]
     pub worktree_id: i64,
     #[sea_orm(primary_key)]
+    pub id: i64,
+    pub is_dir: bool,
     pub path: String,
-    pub language_server_id: i64,
-    pub error_count: i32,
-    pub warning_count: i32,
+    pub inode: i64,
+    pub mtime_seconds: i64,
+    pub mtime_nanos: i32,
+    pub git_status: Option<i64>,
+    pub is_symlink: bool,
+    pub is_ignored: bool,
+    pub is_external: bool,
+    pub is_deleted: bool,
+    pub scan_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
