+++ crates/collab/src/db/tables/worktree_repository_statuses.rs
@@ -1,16 +1,20 @@
-use super::ProjectId;
+use crate::db::ProjectId;
 use sea_orm::entity::prelude::*;
 
 #[derive(Clone, Debug, PartialEq, Eq, DeriveEntityModel)]
-#[sea_orm(table_name = "worktree_settings_files")]
+#[sea_orm(table_name = "worktree_repository_statuses")]
 pub struct Model {
     #[sea_orm(primary_key)]
     pub project_id: ProjectId,
     #[sea_orm(primary_key)]
     pub worktree_id: i64,
     #[sea_orm(primary_key)]
-    pub path: String,
-    pub content: String,
+    pub work_directory_id: i64,
+    #[sea_orm(primary_key)]
+    pub repo_path: String,
+    pub status: i64,
+    pub scan_id: i64,
+    pub is_deleted: bool,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
