+++ crates/collab/src/db/tables/worktree_settings_file.rs
@@ -1,12 +1,16 @@
-use super::ServerId;
+use crate::db::ProjectId;
 use sea_orm::entity::prelude::*;
 
 #[derive(Clone, Debug, PartialEq, Eq, DeriveEntityModel)]
-#[sea_orm(table_name = "servers")]
+#[sea_orm(table_name = "worktree_settings_files")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: ServerId,
-    pub environment: String,
+    pub project_id: ProjectId,
+    #[sea_orm(primary_key)]
+    pub worktree_id: i64,
+    #[sea_orm(primary_key)]
+    pub path: String,
+    pub content: String,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
