--- crates/collab/src/db/worktree_repository_statuses.rs
@@ -1,23 +1,0 @@
-use super::ProjectId;
-use sea_orm::entity::prelude::*;
-
-#[derive(Clone, Debug, PartialEq, Eq, DeriveEntityModel)]
-#[sea_orm(table_name = "worktree_repository_statuses")]
-pub struct Model {
-    #[sea_orm(primary_key)]
-    pub project_id: ProjectId,
-    #[sea_orm(primary_key)]
-    pub worktree_id: i64,
-    #[sea_orm(primary_key)]
-    pub work_directory_id: i64,
-    #[sea_orm(primary_key)]
-    pub repo_path: String,
-    pub status: i64,
-    pub scan_id: i64,
-    pub is_deleted: bool,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {}
-
-impl ActiveModelBehavior for ActiveModel {}
